backend/requirements.txt
asgiref==3.8.1
Django==5.1.1
django-cors-headers==4.4.0
django-environ==0.11.2
djangorestframework==3.15.2
PyJWT==2.9.0
sqlparse==0.5.1
typing_extensions==4.12.2